requirements.txt
PyGithub 
openai 
PyYAML 
python-dotenv 
pandas